fix: refactor login handler
