+++ http/accounts/handler.go
@@ -16,7 +16,6 @@ func NewHandler(router *mux.Router, accUse usecase.AccountUsecase) *Handler {
 
 	router.HandleFunc("/accounts", h.CreateAccount).Methods(http.MethodPost)
 	router.HandleFunc("/accounts", h.ListAll).Methods(http.MethodGet)
-	router.HandleFunc("/login", h.Login).Methods((http.MethodPost))
 	router.HandleFunc("/accounts/{account_id}/balance", h.ListOne).Methods(http.MethodGet)
 	return h
 }
