+++ http/server.go
@@ -7,6 +7,7 @@ import (
 	"github.com/Vivi3008/apiTestGolang/domain/usecases/bill"
 	"github.com/Vivi3008/apiTestGolang/domain/usecases/transfers"
 	"github.com/Vivi3008/apiTestGolang/http/accounts"
+	"github.com/Vivi3008/apiTestGolang/http/auth"
 	"github.com/Vivi3008/apiTestGolang/http/middlewares"
 	"github.com/gorilla/mux"
 )
@@ -19,7 +20,7 @@ type Server struct {
 }
 
 func NewServer(
-	usecaseAcc account.AccountUsecase,
+	accountUc account.AccountUsecase,
 	usecaseTr transfers.TranfersUsecase,
 	usecaseBl bill.BillUsecase,
 ) Server {
@@ -32,7 +33,8 @@ func NewServer(
 	router := mux.NewRouter()
 	routerAuth := router.NewRoute().Subrouter()
 
-	accounts.NewHandler(router, usecaseAcc)
+	accounts.NewHandler(router, accountUc)
+	auth.NewHandler(router, accountUc)
 
 	routerAuth.HandleFunc("/transfers", server.CreateTransfer).Methods((http.MethodPost))
 	routerAuth.HandleFunc("/bills", server.CreateBill).Methods((http.MethodPost))
